refactor(core): simplify Client, extract pipeline and event handlers
