+++ src/vania.ts
@@ -4,6 +4,7 @@ import chalk from 'chalk';
 import { createInterface } from 'readline';
 import { mostrarBannerVania, seleccionarMetodoAuth } from './utils/cli.js';
 import { env } from './config/env.js';
+import { logger, logError } from '@/utils/logger.js';
 
 const SESSION_DIR = env.SESSION_PATH;
 const SESSION_CREDS = `${SESSION_DIR}/creds.json`;
@@ -20,7 +21,7 @@ let firstRestartTime: number | null = null;
 let isAuthenticated = false;
 let shutdownRegistered = false;
 
-console.info(chalk.bold.hex('#FF1493')('\n🦋─ Iniciando VaniaBot IA ─🦋\n'));
+logger.info(chalk.bold.hex('#FF1493')('\n🦋─ Iniciando VaniaBot IA ─🦋\n'));
 
 function hasExistingSession(): boolean {
   return existsSync(SESSION_CREDS);
@@ -34,7 +35,7 @@ function resetRestartCounterIfWindowExpired(): void {
 }
 
 function scheduleRestart(authMode: 'qr' | 'code', delayMs: number): void {
-  console.info(
+  logger.info(
     chalk.cyan(
       `🔄 Reiniciando en ${delayMs / 1000}s... (Intento ${restartCount}/${MAX_QUICK_RESTARTS})`,
     ),
@@ -46,10 +47,8 @@ function scheduleRestart(authMode: 'qr' | 'code', delayMs: number): void {
 }
 
 function scheduleDelayedRestartAfterFlood(authMode: 'qr' | 'code'): void {
-  console.info(chalk.red(`\n Demasiados reinicios (${restartCount}) en poco tiempo`));
-  console.info(
-    chalk.yellow(`⏳ Esperando ${FORCE_RESTART_WAIT_MS / 1000}s antes de reintentar...`),
-  );
+  logger.info(chalk.red(`\n Demasiados reinicios (${restartCount}) en poco tiempo`));
+  logger.info(chalk.yellow(`⏳ Esperando ${FORCE_RESTART_WAIT_MS / 1000}s antes de reintentar...`));
   setTimeout(() => {
     restartCount = 0;
     firstRestartTime = null;
@@ -60,12 +59,12 @@ function scheduleDelayedRestartAfterFlood(authMode: 'qr' | 'code'): void {
 
 function startBot(authMode: 'qr' | 'code'): void {
   if (isRunning) {
-    console.info(chalk.yellow('⚠️ El bot ya está ejecutándose'));
+    logger.info(chalk.yellow('⚠️ El bot ya está ejecutándose'));
     return;
   }
 
   isRunning = true;
-  console.info(chalk.cyan('▶ Iniciando VaniaBot...\n'));
+  logger.info(chalk.cyan('▶ Iniciando VaniaBot...\n'));
 
   childProcess = spawn('node_modules/.bin/tsx', ['src/index.ts'], {
     stdio: 'inherit',
@@ -77,7 +76,7 @@ function startBot(authMode: 'qr' | 'code'): void {
 
   childProcess.on('message', message => {
     if (message === 'ready') {
-      console.info(chalk.green('\n Bot autenticado y listo para operar'));
+      logger.info(chalk.green('\n Bot autenticado y listo para operar'));
       isAuthenticated = true;
       restartCount = 0;
       firstRestartTime = null;
@@ -88,17 +87,17 @@ function startBot(authMode: 'qr' | 'code'): void {
     isRunning = false;
     childProcess = null;
 
-    console.info(
+    logger.info(
       chalk.yellow(`\n⚠️ Proceso finalizado (código: ${code}, señal: ${signal ?? 'ninguna'})`),
     );
 
     if (signal === 'SIGTERM' || signal === 'SIGKILL' || code === 130) {
-      console.info(chalk.green('✓ VaniaBot cerrado correctamente'));
+      logger.info(chalk.green('✓ VaniaBot cerrado correctamente'));
       process.exit(0);
     }
 
     if (code === 0 && isAuthenticated) {
-      console.info(chalk.green('✓ VaniaBot cerrado correctamente'));
+      logger.info(chalk.green('✓ VaniaBot cerrado correctamente'));
       process.exit(0);
     }
 
@@ -117,11 +116,11 @@ function startBot(authMode: 'qr' | 'code'): void {
   });
 
   childProcess.on('error', err => {
-    console.error(chalk.red(' Error en proceso hijo:'), err);
+    logError('Error en proceso hijo:', err);
     isRunning = false;
     childProcess = null;
 
-    console.info(chalk.yellow('⏳ Reintentando en 5 segundos...'));
+    logger.info(chalk.yellow('⏳ Reintentando en 5 segundos...'));
     setTimeout(() => {
       isAuthenticated = false;
       startBot(authMode);
@@ -130,15 +129,15 @@ function startBot(authMode: 'qr' | 'code'): void {
 }
 
 async function gracefulShutdown(signal: string): Promise<void> {
-  console.info(chalk.yellow(`\n⚠️ Recibida señal ${signal}`));
-  console.info(chalk.cyan('🛑 Cerrando VaniaBot de forma segura...'));
+  logger.info(chalk.yellow(`\n⚠️ Recibida señal ${signal}`));
+  logger.info(chalk.cyan('🛑 Cerrando VaniaBot de forma segura...'));
 
   if (childProcess) {
     childProcess.kill('SIGTERM');
 
     await new Promise<void>(resolve => {
       const timeout = setTimeout(() => {
-        console.info(chalk.red('⚠️ Forzando cierre del proceso hijo...'));
+        logger.info(chalk.red('⚠️ Forzando cierre del proceso hijo...'));
         childProcess?.kill('SIGKILL');
         resolve();
       }, 10_000);
@@ -153,10 +152,10 @@ async function gracefulShutdown(signal: string): Promise<void> {
   try {
     if (existsSync(BOOT_FLAG)) unlinkSync(BOOT_FLAG);
   } catch (error) {
-    console.warn('[Vania] Boot flag cleanup error (non-fatal):', error);
+    logger.warn('[Vania] Boot flag cleanup error (non-fatal):', error);
   }
 
-  console.info(chalk.green('✓ VaniaBot cerrado correctamente'));
+  logger.info(chalk.green('✓ VaniaBot cerrado correctamente'));
   process.exit(0);
 }
 
@@ -171,7 +170,7 @@ if (!shutdownRegistered) {
 }
 
 process.on('uncaughtException', err => {
-  console.error(chalk.red(' Error no capturado:'), err);
+  logError('Error no capturado:', err);
   if (childProcess) {
     childProcess.kill('SIGTERM');
     setTimeout(() => {
@@ -184,7 +183,7 @@ process.on('uncaughtException', err => {
 });
 
 process.on('unhandledRejection', reason => {
-  console.error(chalk.red(' Promesa rechazada no manejada:'), reason);
+  logError('Promesa rechazada no manejada:', reason);
   if (childProcess) {
     childProcess.kill('SIGTERM');
     setTimeout(() => {
@@ -199,9 +198,9 @@ process.on('unhandledRejection', reason => {
 async function promptPhoneNumber(): Promise<void> {
   const rl = createInterface({ input: process.stdin, output: process.stdout });
 
-  console.info(chalk.yellow('\nConfiguración de número de teléfono'));
-  console.info(chalk.gray('   Formato: +[código país][número]'));
-  console.info(chalk.gray('   Ejemplo: +529514639799\n'));
+  logger.info(chalk.yellow('\nConfiguración de número de teléfono'));
+  logger.info(chalk.gray('   Formato: +[código país][número]'));
+  logger.info(chalk.gray('   Ejemplo: +529514639799\n'));
 
   const phone = await new Promise<string>(resolve => {
     rl.question(chalk.cyan('➤ Ingresa tu número de WhatsApp: '), answer => resolve(answer.trim()));
@@ -210,22 +209,22 @@ async function promptPhoneNumber(): Promise<void> {
   rl.close();
 
   if (!phone) {
-    console.info(chalk.red(' No se ingresó número. Saliendo...'));
+    logger.info(chalk.red(' No se ingresó número. Saliendo...'));
     process.exit(1);
   }
 
   const cleaned = phone.replace(/\s/g, '');
   if (!/^\+?\d{10,15}$/.test(cleaned)) {
-    console.info(chalk.red(' Formato de número inválido'));
-    console.info(chalk.yellow('   Debe contener entre 10-15 dígitos'));
-    console.info(chalk.yellow('   Puede incluir + al inicio'));
+    logger.info(chalk.red(' Formato de número inválido'));
+    logger.info(chalk.yellow('   Debe contener entre 10-15 dígitos'));
+    logger.info(chalk.yellow('   Puede incluir + al inicio'));
     process.exit(1);
   }
 
   const formatted = cleaned.startsWith('+') ? cleaned : `+${cleaned}`;
   process.env.PHONE_NUMBER = formatted;
 
-  console.info(chalk.green(`✓ Número configurado: ${formatted}\n`));
+  logger.info(chalk.green(`✓ Número configurado: ${formatted}\n`));
 }
 
 async function main(): Promise<void> {
@@ -236,12 +235,12 @@ async function main(): Promise<void> {
 
   if (cliAuthMode === 'qr') {
     selectedAuthMode = 'qr';
-    console.info(chalk.cyan('Usando método: ') + chalk.bold.green('Código QR'));
+    logger.info(chalk.cyan('Usando método: ') + chalk.bold.green('Código QR'));
   } else if (cliAuthMode === 'code') {
     selectedAuthMode = 'code';
-    console.info(chalk.cyan('Usando método: ') + chalk.bold.green('Código de Pareamiento'));
+    logger.info(chalk.cyan('Usando método: ') + chalk.bold.green('Código de Pareamiento'));
   } else if (cliAuthMode) {
-    console.info(
+    logger.info(
       chalk.red(
         `\n Argumento inválido: "${cliAuthMode}"\n\n` +
           'Uso correcto:\n' +
@@ -252,13 +251,13 @@ async function main(): Promise<void> {
     );
     process.exit(1);
   } else if (hasExistingSession()) {
-    console.info(chalk.yellow('⚡ Sesión existente detectada, arrancando directamente...\n'));
+    logger.info(chalk.yellow('⚡ Sesión existente detectada, arrancando directamente...\n'));
     selectedAuthMode = 'qr';
   } else {
     if (!existsSync(BOOT_FLAG)) {
       await mostrarBannerVania();
     } else {
-      console.info(chalk.yellow('⚡ Detectado arranque previo, saltando animación...\n'));
+      logger.info(chalk.yellow('⚡ Detectado arranque previo, saltando animación...\n'));
     }
     selectedAuthMode = await seleccionarMetodoAuth();
   }
@@ -275,6 +274,6 @@ async function main(): Promise<void> {
 }
 
 main().catch(error => {
-  console.error(chalk.red(' Error fatal:'), error);
+  logError('Error fatal:', error);
   process.exit(1);
 });
